✨ feat(users): 닉네임 중복 체크 테스트 파일 추가
